feat: remove Aspire.Host project and related files
