--- docs/Aspire.Host/Aspire.Host.csproj
@@ -1,20 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<PropertyGroup>
-		<OutputType>Exe</OutputType>
-		<TargetFramework>net8.0</TargetFramework>
-		<ImplicitUsings>enable</ImplicitUsings>
-		<Nullable>enable</Nullable>
-		<UserSecretsId>46a3394b-badb-49a7-9c9a-f97ed6969ae3</UserSecretsId>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<PackageReference Include="Aspire.Hosting.AppHost" />
-		<PackageReference Include="Aspire.Hosting.PostgreSQL" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<ProjectReference Include="..\Aspire.MinimalApi\Aspire.MinimalApi.csproj" />
-	</ItemGroup>
-
-</Project>
