+++ src/AspireApp.AppHost/AppHost.cs
@@ -1,9 +1,22 @@
 var builder = DistributedApplication.CreateBuilder(args);
 
-builder.AddProject<Projects.AspireApp_BlazorWeb>("aspireapp-blazorweb");
+// Add PostgreSQL database
+var postgres = builder.AddPostgres("postgres")
+    .WithDataVolume()
+    .WithPgAdmin();
 
-builder.AddProject<Projects.AspireApp_MinimalApi>("aspireapp-minimalapi");
+var productDb = postgres.AddDatabase("productdb");
 
-builder.AddProject<Projects.AspireApp_ApiGateway>("aspireapp-apigateway");
+// Add the Minimal API project
+var api = builder.AddProject<Projects.AspireApp_MinimalApi>("aspireapp-minimalapi")
+    .WithReference(productDb);
+
+// Add the API Gateway
+var gateway = builder.AddProject<Projects.AspireApp_ApiGateway>("aspireapp-apigateway")
+    .WithReference(api);
+
+// Add the Blazor dashboard app
+var blazorApp = builder.AddProject<Projects.AspireApp_BlazorWeb>("aspireapp-blazorweb")
+    .WithReference(gateway);
 
 await builder.Build().RunAsync().ConfigureAwait(false);
